backend/requirements.txt
fastapi==0.115.12
passlib==1.7.4
pydantic==2.11.2
pymongo==4.6.2
python-dotenv==1.1.0
python-jose==3.4.0
uvicorn==0.34.0
argon2_cffi==23.1.0